Modifying geom.prepare_riv() function to ignore columns not given by the user
This makes the function more generalized and allows user to possibly use
it with other geofabrics. However, the functions in the `geom` module
have not been fully tested with other geofabrics and it warrants further
investigation in the near future.

For now, the function is fully compatible with 'MERIT-Basins' geofabric.