Debug/geom.prepare riv (#1)
* initial commit

* initial folders

* messy notebooks as tests for now

* updating notebook to include only one guage - again just tests

* First dirty commit

* Initial hatch package management scheme

* dirty commits

* changing versioning system

* HYDAT initial commit

* correcting first function added

* correcting typehint

* another dirty first commit

* Modifying geom.prepare_riv() function to ignore columns not given by the
user

This makes the function more generalized and allows user to possibly use
it with other geofabrics. However, the functions in the `geom` module
have not been fully tested with other geofabrics and it warrants further
investigation in the near future.

For now, the function is fully compatible with 'MERIT-Basins' geofabric.

* Importing local function for `intersect_topology`

A local function, specifically `find_upstream` was missing for the
`intersect_topology` function. This has been added and the function
works properly.

* Adding the longest_branch function

The function returns a list of river segment IDs of the longest branch
of segments found in a river network. The function uses `networkx` as a
dependency to efficiently investigate the problem.

* Sanity checks completed